requirements.txt
torch
transformers
huggingface_hub
pillow
requests
